Fix memory leak after TextView.setText
When TextView.setText was called on a TextView that the user
previously touched the mInsertionPointCursorController and
mSelectionModifierCursorController was set to null without first
calling ViewTreeObserver.removeOnTouchModeChangeListener. This caused
the ViewTreeObserver.mOnTouchModeChangeListeners to grow when the
user pressed the updated TextView. In some cases this also lead to a
larger memory leak where the TextView held a reference to a
LinearLayout through mParent.